Renamed protected router file
